backend/requirements.txt
python-rtmidi
pychord
sounddevice
numpy
librosa
fastapi
uvicorn[standard]